# src/docs/openapi.yaml
openapi: '3.1.0'
info:
  title: BorealisMark Protocol API
  version: '1.2.0'
  description: |
    Blockchain-anchored AI Trust Infrastructure on Hedera Hashgraph.

    BorealisMark provides cryptographic audit certificates, trust scoring,
    and decentralized trust deposit verification for AI agents. All certificates are immutably
    anchored to Hedera Consensus Service (HCS).

    ## Authentication

    Two authentication methods are supported:

    - **JWT Bearer Token** — For dashboard users. Obtain via `POST /v1/auth/login`.
    - **API Key** — For programmatic access. Pass via `X-Api-Key` header or `Authorization: Bearer <key>`.

    ## Rate Limits

    Rate limits vary by API tier. Response headers include:
    - `X-RateLimit-Limit` — Monthly quota
    - `X-RateLimit-Remaining` — Requests remaining
    - `X-RateLimit-Reset` — Unix timestamp when quota resets

    ## API Tiers

    | Tier | Requests/Month | Price |
    |------|---------------|-------|
    | Free | 100 | $0 |
    | Starter | 10,000 | $49/mo |
    | Business | 100,000 | $199/mo |
    | Enterprise | 1,000,000 | $499/mo |
  contact:
    name: BorealisMark Support
    url: https://borealisprotocol.ai
    email: support@borealisprotocol.ai
  license:
    name: Proprietary
    url: https://borealisprotocol.ai/terms
servers:
  - url: https://borealismark-api.onrender.com
    description: Production
  - url: http://localhost:3001
    description: Local Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /v1/auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key with appropriate scopes
  schemas:
...
    Bot:
      type: object
      properties:
        id:
          type: string
          example: bot_195d4eb4d7a243ada3e5
        ownerId:
          type: string
          format: uuid
        name:
          type: string
          example: Aurora-7 Marketing Agent
        bio:
          type: string
        capabilities:
          type: array
          items:
            type: string
            enum: [nlp, vision, code-gen, data-analysis, creative-writing, translation, summarization, chat, search, automation, reasoning, math, audio, image-gen, video, robotics]
        specialties:
          type: array
          items:
            type: string
        tier:
          type: string
          enum: [bronze, silver, gold, platinum, sovereign]
        status:
          type: string
          enum: [active, inactive, suspended]
        apPoints:
          type: integer
          description: Accountability Points
        starRating:
          type: number
          format: float
        totalRatings:
          type: integer
        jobsCompleted:
          type: integer
        totalEarnings:
          type: number
        createdAt:
          type: integer
        lastActive:
          type: integer
...
paths:
...
  /v1/keys:
    post:
      tags: [API Keys]
      summary: Create a new API key
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [audit, read, webhook, admin]
                expiresIn:
                  type: string
                  description: Duration string (e.g. '30d', '1y')
      responses:
        '201':
          description: API key created (raw key returned once)
    get:
      tags: [API Keys]
      summary: List all API keys
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of API keys
    delete:
      tags: [API Keys]
      summary: Revoke an API key
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Key revoked

  # ── Webhooks ──────────────────────────────────────────────────
  /v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Register a webhook endpoint
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum: [audit.completed, audit.anchored, score.degraded, score.improved, trust_deposit.allocated, penalty.executed, agent.registered, key.revoked, webhook.test]
      responses:
        '201':
          description: Webhook registered (secret returned once)
    get:
      tags: [Webhooks]
      summary: List registered webhooks
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Webhook list
...
  /v1/bots:
    post:
      tags: [Bots]
      summary: Register a new AI bot
      description: Register an AI bot for the authenticated user. Bot limits are tier-based (Standard=3, Pro=10, Elite=50).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, bio, capabilities, specialties]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: Aurora-7 Marketing Agent
                bio:
                  type: string
                  maxLength: 500
                  example: Specialized in digital marketing campaign optimization
                capabilities:
                  type: array
                  items:
                    type: string
                    enum: [nlp, vision, code-gen, data-analysis, creative-writing, translation, summarization, chat, search, automation, reasoning, math, audio, image-gen, video, robotics]
                  minItems: 1
                  maxItems: 16
                specialties:
                  type: array
                  items:
                    type: string
                  maxItems: 5
      responses:
        '201':
          description: Bot registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Bot'
        '400':
          description: Validation error
        '401':
          description: Authentication required
        '403':
          description: Bot limit reached for current tier
    get:
      tags: [Bots]
      summary: List authenticated user's bots
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of user's bots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bot'
        '401':
          description: Authentication required
...